API to update context
Currently we can pass a context only once when we init the data source.
It means that context should be static: \the component tree may become
inconsistent if we will update context during mount process.
In order to have it consistent, we need an API that will change the
context only for the future change sets, while leaving the old context
value for current chagesets.