[CDS] removed comparison operator from CKArrayControllerOutputChange
Comparison for Change doesn't have much sense.
It was used only in one test, which was broken anyway. We were trying to compare order using sets. Lol.
So I have just renamed the test to express it doesn't really test any order and added hashing for `Change` for the tests.