[ComponentKit] Fix animation leaks
Retaining animations on CKComponentController causes a reference cycle:
the component is retained by the animation, which is retained by the
controller, which is retained by the scope handle, which is retained by
the component.

The cycle gets broken when the component is unmounted, but that happens
too late in some cases. This change makes the cycle get broken when the
animation finishes.