Fix CKNetworkImageComponent flash
If we already showed the image in CKNetworkImageComponent, if we download the same URL but the defaultImage is the same, the current code will assume it is a different image and set the defaultImage again and redownload the URL again.
There is no need to do that if the image is already downloaded. All we have to do is update the specifier to make sure if we need to show the defaultImage, we will use the new one when needed.

It is hard to snapshot test this since this is behavior inside CKNetworkComponentView. Since this will require create a new 2nd component and make it use the same view. Happy to create a unit test if this is possible.