Redo how debug "reflow" works
This was never updated after we transitioned away from `CKComponentLifecycleManager`, so it was mostly broken. Now instead of attempting to use the view hierarchy to trigger reflow, we simply message all interested parties and tell them to update themselves.

Crucially this means that offscreen components are recomputed too, not just onscreen ones. Neat.