Add noexcept specification to ComponentKit Core
Presents a guarantee to the compiler that these functions will not
throw. For reviewers' reference, the consequence of violating a
`noexcept` specification is that `std::terminate` is called.

`NSAssert` seems to throw on error, and `CKAssert` wraps `NSAssert`,
so I considered using `noexcept(!CK_ASSERTIONS_ENABLED)` on those
functions that use `CKAssert`. However, local testing revealed that
`NSAssert` failure in a `noexcept` function doesn't result in an
unusable exception backtrace, at least on Mac. (It doesn't seem to
change the exception crash message at all.)

I did a profile build of WildeGuess for simulator (having difficulty
building for device right now) and observed a 1.4% code size
improvement with this diff, which seems to support my hypothesis that
these `noexcept` specifications provide additional optimization
opportunities.