Refactor CKComponentActionAttribute
When this function was originally designed, it was a safe assumption that the number of unique `CKComponentAction` identifiers in the app would be at most ~200. Thus it seemed reasonable to allocate a separate "forwarder" for each identifier and keep the corresponding object in the map forever.

With the introduction of scope-based and block-based actions, this assumption is dangerous; there could potentially be tens of thousands of distinct identifiers, and each distinct action will result in the allocation of a separate forwarder that is essentially leaked forever.

Refactor by keying forwarders on `UIControlEvents` instead of `CKComponentAction::identifier`. This should result in at most ~16 forwarders being active at any given time, regardless of the number of actions.