Ensure that state updates are processed in the order they were enqueued.
Summary: Currently, state updates are held in a `std::unordered_multimap`, which does not guaranteed that items are stored in the order in which they were inserted. As a result, we could be processing state updates out of order, which could results in bad state for components that depend on the order. I've added a unit test to demonstate a situation that processing the state updates could be problematic.

Reviewed By: gkassabli

Differential Revision: D5699440

fbshipit-source-id: d50e6d4bd9db427bb6ae6ff1f1362537f41ad4c9