Remove line that contradicts the steps before it.
In the README Quickstart section, it states in bold: "Before you open any of the Xcode projects in this repo, make sure you run...". Then at the end of the same section it states: "There is no need to run any special commands prior to opening either Xcode project.", which contradicts the previous statement which can lead to confusion.