Remove line that contradicts the steps before it.
Summary:
In the README Quickstart section, it states in bold: "Before you open any of the Xcode projects in this repo, make sure you run...". Then at the end of the same section it states: "There is no need to run any special commands prior to opening either Xcode project.", which contradicts the previous statement which can lead to confusion.
Closes https://github.com/facebook/componentkit/pull/876

Differential Revision: D5949287

Pulled By: lostatseajoshua

fbshipit-source-id: 6efa41b779496fc2d8fd8a7bf286915b4d7a0d41