Make asynchronous preparation of data source modification more like a pipeline
Summary:
Currently, after an async modification is prepared on the bg queue, we dispatch back to main thread, apply the change (slow UI work), then dispatch back to the bg queue to prepare the next modification. Here we are wasting quite some time on the bg queue: 1) dispatch to main queue can be rather slow if main thread is busy 2) UI work done on main thread can easily take tens of ms 3) another dispatch back to bg queue.

This diff proposes a more pipeline like processing approach in which we simply proceed to prepare the next modification on bg queue after the `dispatch_async` to main thread returns. The threading model is arguably more complex and harder to understand, but if we can gain some measurable speedup, it might be worth it.

To minimize impact on existing code path, I used a `NSLock` and only initialize it if the pipeline option is enabled

Reviewed By: ocrickard

Differential Revision: D6011849

fbshipit-source-id: 9074575adfdbc927290952fad522d65f2f85b3a7