Expose border width in CKFlexboxComponent
Summary:
What:
Expose borderWidth as a property in CKFlexboxComponent

Why:
Yoga uses the border-box sizing model. This means that if I have a 100x100 flexbox with a 10pt border, the size of the flexbox *including* the border should be 100x100 - so the inner contents should be shrunk.

This doesn't work at the moment. E.g. if I have a component that draws a border around a flexbox, it shouldn't be managing the size. It should pass the border width through to Yoga, which will size the flexbox appropriately.

Reviewed By: gkassabli

Differential Revision: D6142002

fbshipit-source-id: 089af7e1f13c97e5e238233596377f8d96a2c955