All negative points in CKRelativeDimension
Summary:
For profile picture components, we would like to offset the badge such that the center of the badge lies on the circumference of the profile picture. This offset lies in the range of -0.6pt to -1.1pt for badges of size 40pt to 64pt. Unfortunately, since negative CKRelativeDimensions are not allowed, this ends up crashing. Several questions and thoughts:

1) What was the basis for this assert? Are there instances where a negative offset will break certain assumptions in componentkit layout? Removing these asserts doesn't seem to cause problems and the component renders successfully. For context, [FBBadgedProfilePictureLayoutComponent](diffusion/FBS/browse/master/fbobjc/Libraries/FBComponentUI/FBBadgedProfilePictureLayoutComponent/FBBadgedProfilePictureLayoutComponent.mm;74fbadf09c8b60e22a2f1cd226eb864c4203d8aa$48) is where I want to apply this negative offset.
2) The alternative to not using a negative offset will lead to really ugly layout code elsewhere. For instance, if I want the body text of a list cell to be laid out 8pt to the right of a profile, it now has to be 8pt - offset. This logic will also cause complications for vertically centering the profile picture component. Furthermore, callers will have to have different layout logic for badged/non-badged cases.

Ideas on how I can work around this? Is removing the assert or having an option to bypass it a good idea?

Reviewed By: gkassabli

Differential Revision: D6289059

fbshipit-source-id: 8c440153f16edc48e6c0927142c2f6cff4ddb070