adding type checking for structs in action send
Summary:
We noticed there is an issue when sending c++ types in CKAction (mainly happens with GQL) where types are "slightly" mismatching due to weird c++ stuff. The types are actually the same and everything works other than the assert being thrown.
Since c++ has no way of retrieving a type from a name at runtime (only to opposite), this is sort of a hack - based on Apple's documentation on encoding of struct types.
https://developer.apple.com/library/content/documentation/Cocoa/Conceptual/ObjCRuntimeGuide/Articles/ocrtTypeEncodings.html

Reviewed By: ocrickard

Differential Revision: D6447108

fbshipit-source-id: 7db0cd48402cec8db27bb6a0a112d1c2b86f04c1