Add the concept of component cache to CKComponentLayout
Summary:
Add the option to build a component cache in the root layout when computing the root layout.

The cache will contain a map from components that has a component controller to their layout.

This is an optional feature; the default mode will be without a cache.
The cache will be stored in the extra dictionary in the root layout struct and will accessible  through a static method.

Reviewed By: gkassabli

Differential Revision: D6812059

fbshipit-source-id: 2f00b09a258953010dba89b479ca0aef54544804