Add RTL support in Flexbox
Summary:
This takes advantage of RTL support in Yoga (https://yogalayout.com/docs/layout-direction) and enables it in `CKFlexboxComponent`.

RTL is enabled by default, meaning if your application is rendering in RTL, `CKFlexboxComponent` will automatically adjust to layout in RTL as well. For cases where you don't want the automatic behavior, you can specify a direction via the `layoutDirection` option in `CKFlexboxComponentStyle`.

This diff uses a `NSParagraphStyle` method to get the application's layout direction. Normally, you can do this via a call to `[[UIApplication sharedApplication] userInterfaceLayoutDirection]`. However, in Xcode 9, this fires a run time warning from the main thread checker and since ComponentKit does layouts in the background, it will fire the first time flexbox tries to use this as well.

So instead of directly calling `userInterfaceLayoutDirection`, this calls `[NSParagraphStyle defaultWritingDirectionForLanguage:nil]`, which is what `userInterfaceLayoutDirection` is calling under the hood.

Passing `nil` to that method causes it to internally call `[NSParagraphStyle _defaultWritingDirection]`, which then uses a chain of Core Foundation functions, (`CFBundleCopyPreferredLocalizationsFromArray(CFBundleCopyBundleLocalizations(CFBundleGetMainBundle()))`) to get the writing direction of the preferred localization.

Reviewed By: Andrey-Mishanin

Differential Revision: D7087003

fbshipit-source-id: 55d6116b3ac83360860a5529796e6f6e0e50b37f