Unified Render and Layout - Introducing CKRenderWithSizeSpecComponent
Summary:
From the root component it calls `buildComponentTree:...` and `render:` until we reach a `CKRenderWithSizeSpecComponent`, then it starts layouting.
While layouting as soon as it reaches a CKRenderWithSizeSpecComponent the framework will call
```
computeLayoutThatFits:state: (or the 4 parameters version)
```
that needs to be overriden by the user. In this method every child of the component is attached to the component tree and layouted (those operations are transparent to the user, all of them happens when they call
```
[self measureChild:_childComponent ...];
```

Pro:
- Simpler API
- Components can be created based on what you need and not just chosen at layout time
- Components created during layout can have a state

Cons:
- I will not work if we have a CKComponent subclass (not CKCompositeComponent, or CKFlexbox.. just plain CKComponent) in the hierarchy
- CKComponentHostingView will always generate and layout the component (instead of just layouting in some cases)

Reviewed By: kfirapps

Differential Revision: D7514923

fbshipit-source-id: a4d681d84809a525c5efb67190b79d5619d2fb1f