Add forceParent parameter to buildComponentTree method
Summary:
Today each render component can either be an owner component or not. Based on this definition we create a corresponding `CKOwnerTreeNode` or `CKRenderTreeNode/CKTreeNode`.

In order to explore the option to move to a parent based component key, we are adding this param. It will be passed by `CKBuildComponent` to all the component in the tree; according to this param the component will choose what type of tree node to create.

Reviewed By: aCorrado

Differential Revision: D7790896

fbshipit-source-id: eeb9edb61d27e262e58c936076f206ef2f52a3d9