Collecting animations from components
Summary:
After determining which components need to be animated with `CK::animatedComponentsBetweenScopeRoots()`, we collect animations into an instance of `CK::ComponentAnimations` by calling the new `CK::animationsForComponents()`.

The implementation maps each animated component to a vector of animations by calling `animationsOnInitalMount` or `animationsFromPreviousComponent:` on `appearedComponents` or `updatedComponents` respectively. The resulting vector of vectors is then flattened to a one-dimensional vector of animations by calling `CK::Collection::flatten()`.

In order to be able to compare animations in tests I've added `CKComponentAnimation::isIdenticalTo()` which returns `true` if two animations have have the exact same hooks (basically, this means that one animation struct is a bitwise copy of another). I'd love to have this method to be confined only to the test target, but it needs to access the private `hooks` field in `CKComponentAnimation`.

Reviewed By: kfirapps

Differential Revision: D8185921

fbshipit-source-id: c50a0e5e97beb8734e0beff14f37239994457e3a