Allow taps to pass through the hosting and root layers
Summary: In order to better allow components to be layered over views, I have added an option to both the component hosting view and the component root view which allows taps to pass through them. This allows the hosting view to act as a mere container and be layered over large portions of the screen and still allow taps that occur in views behind the component layer to still work.

Reviewed By: kfirapps

Differential Revision: D8232854

fbshipit-source-id: da19b8a208472f6e4f6dbdfc34e7b19751de873e