Fix search issue crashing sonar
Summary:
Some components have responders with variant "Block".
When you try to get the initialTarget of such a responder, we get an assertion failure because this is never supposed to happen. We do this because in sonar, we go about getting all the information we can about everything.

This is rare, but the bug gets surfaced whenever you search for "a" on facebook for iOS in news feed on sonar.

To get around it, whenever we see a responder of this type, I'm omitting it from the response.

Reviewed By: danielbuechele

Differential Revision: D8351783

fbshipit-source-id: 7bfba06d620614960edbc442c60e0815480210a8