No longer collecting animations at build stage
Summary:
Animations need to be collected at mount stage otherwise we won't be honouring the `animationsOnInitialMount` API contract.

Imagine a situation like this:

1. N items are added to a data source.
2. Component trees are created for all added items regardless of whether they are on-screen or off-screen.
3. For all new components, we would collect corresponding appear animations.
4. Collected animations are applied at mount stage for on-screen components only.
5. A changeset is applied that updated off-screen items, so we rebuild the corresponding component trees.
6. For all updated components, we would collect their *change* animations.
7. When the off-screen components finally become visible, we'd run *change* animations for them, not appear animations, thus breaking *initialMount...* part of the contract.

The solution I'm proposing is just moving tree diffing and animation collection to mount stage, so we are comparing not just the current scope tree and the previous scope tree, but the current scope tree before we *mount* it and the previous *mounted* scope tree. This will be implemented in the next diffs.

Reviewed By: kfirapps

Differential Revision: D8480816

fbshipit-source-id: f363b6c087d16c9f36d8df1060c806dfd09d6fab