Improved changeset verification for moves
Summary:
The original problem in verification algorithm was that we were validating moved index paths against the final state of the data source. This is incorrect because "from" index paths in moves are in original index coordinate space, before any removes have been applied. This means that they need to be validated against original section indexes.

Moreover, the verification algorithm also uses these indexes to verify a number of elements in section, so just changing it to compare "from" index paths against original section counts is not enough. We also need to increment and decrement the number of items in correct section, that is, we need to decrement the number of items in `originalSectionCounts` using the section index *before* applying deletions, but increment the number of items in final `sectionCounts` using the index of the same logical section but *after* applying deletions. In my previous attempt to fix this problem I forgot to adjust the section index I'm using to increment the number of items, hence we were having false positives. So, to make these index adjustments I'm using `CK::IndexTransforms` I added in previous diffs.

Reviewed By: fabiomassimo

Differential Revision: D9042828

fbshipit-source-id: b67662d1a2e710143dc8ae60175087b65a5564d1