Use free function instead of method
Summary: Type encodings for all ObjC method signatures are stored as literal strings in the binary. If one of the argument types is something that transitively includes C++ templates, it's type encoding size will baloon very quickly (see P59826536 and [this SO answer](https://stackoverflow.com/questions/11527385/how-are-the-digits-in-objc-method-type-encoding-calculated/11527925#11527925)). Free functions do not have this problem.

Reviewed By: kfirapps

Differential Revision: D9082811

fbshipit-source-id: f37e74c125a84e4c5febcb92e303ae429d3ddc5d