Fix retain cycle in optimistic view mutation teardown
Summary:
This appears to cause a retain cycle wherever an optimistic mutation is applied.

Current ownership looks like:

view [owns] mutationTeardowns [owns] view

So we want the mutation teardown to hold a weak reference to the view or, as is the case here, let someone else pass in a reference to the view.

Differential Revision: D9171774

fbshipit-source-id: 8c44cfe47ae6d0dc615415d7e23ecb1ade90dca6