Use the new tree node in case of a fatser/props state update
Summary: Before this change, in case if reusing the previous component, we used also its previous treeNode which was linked to the old component. Now, we use the new component, but link it to the child and the tree node of the previous component. This way, we can guarantee that the parent of this component holds the correct component that we are going to use during render time.

Reviewed By: kevin0571

Differential Revision: D9380646

fbshipit-source-id: 2dbfeb2691f728dfdf4e80943dc3af15303b1eb7