Set data source ID on the correct queue
Summary: The data source ID (used to verify that methods were being called on the right queue) was being specified on the wrong queue: it should be set on the passed-in `workQueue` rather than the internal serial queue that targets the passed-in `workQueue. The reason for this is that `dispatch_get_specific` will return a context value if the key is defined on the specified queue **or** it's target queue. By setting the data source ID on the internal queue, the assertion would fail if the public methods were called on the `workQueue` that was passed in.

Reviewed By: kevin0571

Differential Revision: D10210759

fbshipit-source-id: 1a98c477556b9e042318bff180767fcd0cfc58aa