Add CKComponentHostingView options for scope root and pending state updates
Summary: This is for the use case where we initially generate the component outside the component hosting view (e.g. on a background queue, prior to the view/view controller being created) and we want to transfer ownership of the scope root to the component hosting view once it is available. See code documentation for more details.

Reviewed By: ksenks

Differential Revision: D10151152

fbshipit-source-id: 14d3c0ff4911e2fb2711372ff89cd46865245acc