Take contentOffset into account when splitting update changeset
Summary: The logic of the previous implementation always assumed that the scroll view was at the top. This assumption does not hold when we have state updates that happen when you're scrolling down (e.g. tail insertions), where the update changesets won't be split correctly. This adds a `CKDataSource` API to inform it of the content offset when applying changesets.

Reviewed By: ksenks

Differential Revision: D13391005

fbshipit-source-id: f163d32496705b412261acdce7321d09d49c0556