Pass viewport size information to data source
Summary:
This change has no impact unless changeset splitting is enabled.

When changeset splitting is enabled, we currently assume that the size of the viewport is the same estimated size that was specified in `splitChangesetOptions`. The real size at the time a changeset is actually being applied could be different, and the changeset splitting will be innacurate if there is a difference between the initially specified estimated size and the actual size. This adds an API to `CKDataSource` that allows for explicitly specifying the **current** viewport size, which will override the estimated size from the configuration if specified.

Reviewed By: ksenks

Differential Revision: D13444218

fbshipit-source-id: fdc7d3d93066a1f7434f8936a5fa7d3023fe4ed4