Setting custom baseline function to all places where CkFlexboxAlignItemBaseline is used
Summary:
Added a custom baseline function which returns the height of node in all the places which use CKFlexboxAlignItemsBaseline so that old behaviour does not change

Context:
Stack of diff is to make the behaviour of baseline consistent with web standard where default baseline of node is baseline of first child of node.
Currently in CK baseline function is set on every yoga node and in baseline function we check if custom baseline is being set or not if yes we simply return that value else we return the height of node.
In this stack we are now setting baseline function only when it users of CK set a custom baseline function + to support old use cases of custom baseline value.

Reviewed By: alickbass

Differential Revision: D13783329

fbshipit-source-id: 09e696c05655b188b5873c4c5fb9de2ac61f160e