Declare conformance to CKComponentStateListener in CKDataSourceProtocol
Summary: When using the seeded data source state API, we need to be able to generate the data source items with an alternative "proxy" state listener that can then be redirected to the real data source that is being seeded. In order to apply the queued up state updates to the new data source, we need to know that it conforms to the `CKComponentStateListener` protocol. Therefore, make this protocol conformance public in `CKDataSourceProtocol`.

Reviewed By: ksenks

Differential Revision: D14295086

fbshipit-source-id: 16dc4d871611aca58ebda5ecf0ad8abd268a18b1