Fix issue with moves to just inserted section
Summary:
Moving exisiting items to a just inserted section is totally fine, but the validation was crashing instead.

If the destination section of a move wasn't present in the previous data source state, we need to increment to new section counts instead.

However, if the move is happening within the section, the destination index should always be validated against the new section counts.

Reviewed By: kevin0571

Differential Revision: D14477544

fbshipit-source-id: 304398ee6de319cdd3e0349bdd1bdb9acda82ba7