Add changeset modification generator delegate to ck data source
Summary:
Currently when we're wanting to tweak the changesets which
get generated by the CKDataSource we're having to change the code
within it. For example we currently internally are forking around
if we split changesets or not based on composition. It would be
nicer to be able to do this by composition so that as the number
of ways we want to configure and test things increases we don't
clutter up the CKDataSource code. This diff adds a delegate to
do just that.

Reviewed By: kfirapps

Differential Revision: D14572323

fbshipit-source-id: 10284a8d0c060e56e57f3fa508d79e1841ff7600