Avoid component reuse on reload
Summary:
Pass `ignoreComponentReuseOptimizations = YES` on reload.

The idea is that during reload, we want to recreate all the components in tree. This mainly going to affect the use case of `CKComponentConstContext`, which allows us to optimize the component creation while using component context. However, in some case, we would like to avoid these optimizations. For example in the use case of passing a theme in `CKComponentConstContext`. In 99% of the times, we don't need to worry about a value change. However, in some case, we do. Reload will be the answer here as it's going to ignore component reuse.

Reviewed By: kevin0571

Differential Revision: D14663917

fbshipit-source-id: 5d5c2c03b7fad25d02adbe832e6547d3be569ce7