Add support for rate limiting component generation in parallel row layout
Summary:
Currently in the parallel row layout world we often starve our component generation
threads by having more layouts going than there are number of cores on the device. This diff
attempts to ensure that this will never happen by rate limiting the number of components we're
generating at a time to the number of cores - 1 (leaving a core for the main thread).

This also introduces a race where we might have the main thread available to do the work before
we've actually scheduled the item for computation. To fix this we allow the main thread itself
to do the work in that case.

Differential Revision: D15839036

fbshipit-source-id: 2c5a307d97fce23d0331f64ff61e4b3a82144b95