Make parallel row layout support sorting updates
Summary:
In the current world of parallel row layout when an update
is happening above insertions all of the insertions end up blocking
on the main thread for the update to be generated first. Instead
we want to order the updates to be first so that we're more likely
to get full parallelism between mounting and generating components.

Differential Revision: D15839039

fbshipit-source-id: 12bef09b3bacba6fc22df6c20d7e5b929bc7feac