Proper duration calculation for parallel groups
Summary:
If an animation with an automatic duration like sequence is being combined with another animation in a parallel group, the duration of this group can't be set externally. Instead, it should be calculated as a maximum of the combined animation durations.

However, if a parallel group combines two animations for which durations *can* be set externally, it should use the same logic as before (propagate the explicitly set duration if it's present).

This calculation is now moved to to a separate function template `durationForParallelGroup()` that has overloads for both cases.

Reviewed By: kevin0571

Differential Revision: D16207120

fbshipit-source-id: d70b51d758dfb22b5a5e273ea508af7f895b5874