Make Optional conditionally trivially copyable
Summary:
Make sure if `Optional` is wrapping a type `T` that is trivially copyable, the whole `Optional<T>` is trivially copyable too.

Basically, now `Optional` has all its copy / move constructors / assignments set to default, which means the actual implementation of these methods is delegated to the `Storage`.

We now select the implementation for `Storage` based on whether `T` is trivially copyable or not. If it is, we'll select the `Storage` implementation that also has all its constructors defaulted. Otherwise, we select the implementation that does what constructors in `Optional` were previously doing.

Reviewed By: kevin0571

Differential Revision: D16378301

fbshipit-source-id: dd5ff7913dea6562dffdcf99e38aa287d52ba036