Non-null pointer type
Summary:
`CK::RelaxedNonNull<Ptr>` has all the features described below but allows *implicit* construction from a nullable pointer. This type is only supposed to be used in client-facing APIs such as `updateState:mode:` or `CKComponentAnimation` so client code doesn’t get littered with calls to `CK::makeNonNull`.

The second type, `CK::NonNull` inherits from `CK::RelaxedNonNull` but only allows *explicit* construction from a nullable pointer. This is the type that should be used for the internal APIs of the framework.

### Features

* No default constructor
* Call ObjC methods as usual
* Access ivars as usual
* Conversion back to nullable
* Comparison to nullable

### Compile-time Checks

* No construction or assignment from `nil`
* No comparison to `nil`
* No conversion to `bool`

Reviewed By: kfirapps

Differential Revision: D16620674

fbshipit-source-id: 7f41518c8c9f984b7656c2ddb416b6fb465dc80a