Return last mounted component when `latestComponent` is used
Summary: We should return last mounted component if it's still alive instead of latest component. This avoids the case where state of datasource is not yet applied and component is requested to be mounted again. In that case, returning last mounted component represents the component that is currently being mounted.

Reviewed By: kfirapps

Differential Revision: D16712943

fbshipit-source-id: 17c45a996a1323c1f4ec892e2b1c78cb54139b8a