Removal of enum for PropId
Summary:
We're not using an enum for the `PropId` - it's now a namespace with `static constexpr auto` values.

Modifying `ComponentBuilder.h` thus now doing manual bitmask operations instead.

We're now fixing the type of the non-type template parameter to `PropsBitmapType` i.e `unsigned long long` which is reducing the amount of template type parameters we have to pass around

Reviewed By: Andrey-Mishanin

Differential Revision: D17741714

fbshipit-source-id: 95ed12d36d57d73320c9da20c78a73d16ffa58fa