Adds `didInit` lifecycle event to `CKComponentController`
Summary: Adding a `-didInit` lifecycle event on `CKComponentController`. This event is always trigerred on the main thread and symetrical to `-invalidateController`. Clients should favour using `-didInit` instead of overriding `-initWithComponent:` since the latter may not be called on the main thread *and* doesn't enforce that its component won't get discarded.

Reviewed By: kevin0571

Differential Revision: D18477662

fbshipit-source-id: 0337ff02ef89e46861e56d17066fad85035dab9a