Variant
Summary:
`CK::Variant` is useful for representing a set of mutually exclusive alternatives represented by instances of `Types` . For example:

```
 CK::Variant<int, double, std::string>
```

can store either an `int` , a `double` or a `std::string`, but only one of them at a time.

In order to access the currently active alternative, a set of `match()` member functions is provided, e.g.

```
 CK::Variant<int, double, std::string> v = ...;
 v.match(
   [](int i) { // Process int value },
   [](double d) { // Process double value },
   [](const std::string &s) { // Process string value }
 );
```

Reviewed By: cuva

Differential Revision: D18622810

fbshipit-source-id: 44a2daec64bf20f9e533e5cfbd21ecc1ec27c7f1