Main thread affined assocaited object map
Summary:
Since objc's associated object map is shared in the entire runtime, it would be great to test out a CK scope associated object map to avoid global spin lock.
As a start, we can replace call sites that are main thread affined with a lock-less CK associated object map.
NOTE: we still need to pay for the cost of using `objc_setAssociatedObject` for the first time when an object associates with value because we need to get notified when this object is deallocated.

Reviewed By: Andrey-Mishanin

Differential Revision: D19463166

fbshipit-source-id: 036cec58bcb8994b377921fee53cd2c0ceff0142