Erasing hosting view type in delegate method
Summary: `CKComponentHostingView` gets imported into Swift as a generic class with type parameters for the model and the context. As a result, the delegate method has to specify the type parameters as well, which is unnecessary complicated. Work around this by declaring the parameter of the delegate method as just `UIView`.

Reviewed By: cuva

Differential Revision: D19925811

fbshipit-source-id: 62cdea5ec827aa7647f152b3f4ef8d08b1d711da