Add unique id to inspectable views
Summary:
We need a way to uniquely identify components when there are several identical component trees on the screen. This is mostly relevant for collection / table views, but I think it might be a good idea to properly handle hosting views as well.

Both the hosting view and the root view conform to `CKInspectableView` and both need this functionality, so it made sense to introduce the unique identifier method as a part of `CKInspectableView`.

Reviewed By: kfirapps

Differential Revision: D19949914

fbshipit-source-id: b276792048cd23e469bdf7eabf9f9bc523a3f3a5