Migrate from component class to component type name
Summary:
The component class is often used to identify the type of component. As we're working to make components less relyant on classes I'm migrating from `Class` to `const char *`.

When component types need to be compared only a pointer comparison is performed. Therefore, `typeName` must return a constant static string.

Reviewed By: kfirapps

Differential Revision: D20180484

fbshipit-source-id: 138bca32cdb36eef80e82b8a10b7ba963cbb8fae