Add support for Optimistic View Mutations with minimum persist time
Summary:
Optimistic view mutations "teardown" (revert their changes) at the very next mount with expectation that the next mount would contain the correct state to make the optimistic mutation permanent. This assumption does not always pan out and, infact, the teardown is non-deterministic from the caller/mutator's point of view.

Optimistic mutations work as expected as long as no other state-updates occur at approximately same time. This can cause the "flicker" bugs to randomly regress as (supposedly) unrelated bits of code are added or changed. What happens is that other unrelated state updates have already queued a mount causing the optimistic mutation to be prematurely reverted. When the expected state-update does occur, it causes another mount. These two mounts are on different iterations of the runloop. Visually, this series of two independent mounts results is a flicker.

This diff adds two `CKPerformOptimisticViewMutation` overloads that take an explicit `CFTimeInterval persistTime` param. The param defines how long the optimistic mutation should stay in place after which time the mutation will be automatically reverted. The mutations (and their teardown) are both deterministic in the sense that they are guaranteed to be applied for at least the specified time and will generally be torn down at the same time (via `dispatch_after`).

Because a mutation can persist longer than the next mount, the state may now change many times before teardown. This means the reverted/torn-down state cannot simply be state at the time the optimistic mutation was applied. Similarly, multiple mutations for the same effective property with different timeouts also mean the reverted state could be in-fact be many possible states:

* The original state before the optimistic mutation
* The state of another optimistic mutation that has yet to time out
* A new state that has been fully applied and mounted since the optimistic mutation

The new optimistic view mutations with `persistTime` will correctly restore to the logically expected state. State will be reverted to any one of the above states and when two or more other optimistic mutations exist, the state of the most recently applied optimistic view mutation will be applied (and if that has already timed-out then it would be the state of the next most recent and so on).

> The state after an optimistic mutation has been torn down will be the expected state if the optimistic mutation had not been applied in the first place - taking into account all other state changes and (still alive) optimistic mutations that were applied before or since.

If a persistTime of 0 seconds is specified then the behaviour will match classic optimistic view mutations.

Reviewed By: cuva

Differential Revision: D21058289

fbshipit-source-id: 287e62ec4c06b4941df17b2f5c33fb99839066fb