Remove CKIsRunningInTest
Summary:
We don't need this; `CKComponentTestRootScope` is intended for use in unit tests.

Looking up classes by name to special case tests is an anti-pattern, in my opinion.

Additionally, wrap all macros in `do {} while (0)`. This is macro best practice, otherwise you can end up with weird edge cases:

```
if (true)
  CKValidateComponentCreation();
else
  DoSomethingElse();
```

The previous implementation of the macro would have attached the `else` to the macro's `if` statement! Scary stuff.

Reviewed By: kevin0571

Differential Revision: D21242509

fbshipit-source-id: fa992e5fa06a73337db35d7ea4619c84e6c956fb