Remove conditional for deep yoga trees
Summary: The fix for Composite component and Overlay layout component is already landed not only for NT but for entire codebase, so this branching is not necessary. More over, we want to propely set the percent on the node itself and not resolve it manually, as at that given moment there might be no flex prent width available and this the percentage will be calculated incorrectly. See test plan for example

Reviewed By: fabiomassimo

Differential Revision: D21685656

fbshipit-source-id: ef4c44f759ac50a76386ef0beb52778438a3e02c