CKPerformMount & CKPerformUnmount
Summary:
Now that `CKViewConfiguration` no longer uses templates, we can hide this implementation and expose only a function signature.

Additionally, we can dispense with `CKMountController` as a concept entirely. This involves the somewhat odd model of having a `unique_ptr` to a mount controller, which itself holds a `unique_ptr` to `CKMountInfo`.

Instead, we just have two helper functions: `CKPerformMount` and `CKPerformUnmount`. These can be used by `CKMountable` implementations to do mounting and unmounting in a standard way.

A follow-up goal is to migrate `CKComponent` to use these helper functions.

Reviewed By: kfirapps

Differential Revision: D22358839

fbshipit-source-id: 0c349bc999da48bea210d423523bcc35605aee0c