Change `_scopeIdentifierAndResponderGenerator` to `_scopedResponderAndKey`
Summary:
Replacing `std::pair<CKScopedResponderUniqueIdentifier, CKResponderGenerationBlock>` with a struct containing both the scope responder (which used to be captured by the generator block) and the responder key (which also used to be captured by the responder block).

We don't really need the block there - and as we've migrated towards using the `scopedResponde.uniqueIdentifier` instead of `handle.globalIdentifier` to identify the target component we can do by simply keeping a ref to the responder and the key.

Reviewed By: kfirapps

Differential Revision: D22136042

fbshipit-source-id: 212a5c1d20e63dae0f3eb6ac1327bed2311849c5