Adds `isEmpty` flag on scope root
Summary:
Last week I added an assert in `CKBuildComponentTrigger` in case `scopeRoot.rootComponent == nil && stateUpdate.empty()` which resulted in T76066397.

To remedy this (i.e a nil `rootComponent` but a state update) I've added a flag on the scope root which is initially `YES` and `NO` when recreated. This should be a better indicator on wether it's a new tree being built.

Reviewed By: kfirapps

Differential Revision: D23840287

fbshipit-source-id: 1791b99ccf3d0ed1fa70b1a3476f79b0e6d50a45