Adds protocol `ViewConfigurationRepresentable`
Summary:
Adds a way for views to describe their view. This protocol is opt-in.

Wether `ViewAttributeAssignable` is required or not is up for debate as typically attributes are passed to a `View` initialiser. However Swift values give use the flexibility to copy and modify which this leverages.

Example:

```
let v = ComponentView<UIImageView> {
  (\.image, image)
}

let vWithRedBackground = v.backgroundColor(.red)

return Flexbox {
  v
  vWithRedBackground
}
```

Reviewed By: Andrey-Mishanin

Differential Revision: D23838452

fbshipit-source-id: e42f0778afe82e5e89dba5efba4e9c0bca111c9a