Cleanup unnecessary `Inflatable` generic argument
Summary:
Using a constrained generic doesn't seem necessary as the right overload of `inflateComponent(with:)` is always called and `ComponentInflatable` doesn't have an associated type (unlike `View`, for example). I was under the impression it was needed for the type system to be able to pick the right `inflateComponent(with:)` implementation but that's not what I'm observing now.

When it comes to modifiers - I can't see a reason (for now) to expose the type it wraps. In the future this might be needed to place constraints on `Views` being accepted as props for example (while still allowing to pass `MyView` or `LayoutModifier<MyView>`). Let's keep it simple for now.

Reviewed By: kevin0571

Differential Revision: D24757493

fbshipit-source-id: 8af1ba29ad892c07bf944df1f88376a69ed0a79a