CKComponentStateListener optimization
Summary:
`CKComponentScopeHandle` always performs thread hop when processing an update which might be suboptimal for certain listeners. Adding a class method on `CKComponentStateListener` which allows to tune this behavior.

Note: ideally, I'd let listeners to decide if they need to process updates on the calling thread; however, this would require a lot of changes on the client side.

Reviewed By: kevin0571

Differential Revision: D24852325

fbshipit-source-id: 26a8343f37c0a44c0af687cde555f1c650b56f24