Adds a read-only property wrapper `ViewModel`
Summary:
Unlike the `State` property wrapper, `ViewModel` is read-only. While its reference is still stored in the component's state, it allows creating and owning a class which itself owns some mutable state.

This is somewhat akin to `StateObject` in SwiftUI, minus the `ObservableObject` requirement which comes from `Combine`.

`ViewModel` allows to interface with delegates (since the view model is a `class`). `ViewModelState` will be added in a subsequent diff to permit passing data back to the view.

Reviewed By: kevin0571

Differential Revision: D26483712

fbshipit-source-id: 9ae12b4e0d8b3b6a18746ef2eb7a66822048b974