Adds a `@ViewModelState` property wrapper
Summary:
`ViewModelState` allows storing some state in a `ViewModel` (somewhat aking to `Published` in SwiftUI minus the combine dependency) which when modified triggers a state update on the component owning the `ViewModel`.

This is a very basic way of delegating changes back from the view model to the view sort of automagically (since `View` are structs, they can't nor should be the delegate of anything).

Upon changing a `ViewModelState`, a no-op state update is enqueued which will cause the tree to be reflowed and the new value to be read from the `viewModel` (itself owned by the component's state).

Reviewed By: kevin0571

Differential Revision: D26483708

fbshipit-source-id: 8e395a0839f620314f22945aba2760290fccd365