Pass actual component class when linking the scope handle.
Summary:
Since we've added `SwiftReusableLeafComponent` and `SwiftReusableComponent` for collection support, the assumption that the concrete component class will always be `SwiftComponent<View>.self` doesn't hold.

Not a huge fan of the repetition between the two but there are a couple of things that currently make this hard to refactor into the class initialiser for now.

Reviewed By: kevin0571

Differential Revision: D26690079

fbshipit-source-id: 30e3ac6e1bdcf580b210dd4d5b8e88300e1db1b6