Fix race condition by caching value on `TreeNodeValueStore`
Summary:
This diff fixes a race condition if a state update is done on a node while a new generation is being calculated (whose trigger wasn't a state update from this specific node).

When performing a state update two things need to happen:

* The state update needs to be enqueue through the update state block for the new state value to be computed when a new tree gets regen.
* A local cache of the state needs to be updated as otherwise reading the state value right after modification would yield the previous value (as the state property wrapper now offers a synchonous API as opposed to the previously async `-updateState:`).

I had initially decided to change the value on the tree node and avoiding the race by capturing a copy of the vector so that subsequent state updates wouldn't affect what gets read on a background thread while a new tree was being generated. This indeed solved the race condition if we were updating a state of a node _which already has a state update enqueued while another generation was being calculated - but not if the sunsequent tree generation was caused by something else (props update or another node's state update).

TL;DR: The state value is now cached on the property wrapper store (which is a reference type) and requires to be read on the building thread or main thread and only updated on the main thread (post build phase). This should be sufficient to mitigate any race conditions there.

Reviewed By: Andrey-Mishanin

Differential Revision: D27008559

fbshipit-source-id: 01bd8771307b4655a6efa59365b434fec93040ba